% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{genAFold}
\alias{genAFold}
\title{Calculate parameters for differential expression test base on absolute counts differences}
\usage{
genAFold(nncounts, cond, preval = 0.05, qforGeneralSD = 0.75)
}
\arguments{
\item{nncounts}{matrix for read count.}

\item{cond}{factor for condition.}

\item{preval}{pre-defined uncertainty level as control to penalize absolute count difference, default is 0.05.}

\item{qforGeneralSD}{quantile for estimating general SD of fold-change, default is 0.75.}
}
\value{
A list with log2 foldchange and general SD for calculating pvalue
}
\description{
Calculate aFold for each gene and general sd
}
\details{
shifted and calculate a set of parameters from normalized counts table before \code{\link{callDEs}}
}
\note{
This function should run after \code{\link{normalFactors}}.
}
\examples{

data(simuN5)
obj <- ABSDataSet(counts=simuN5$counts, groups=factor(simuN5$groups))
mtx <- counts(obj,TRUE)
aFold <- genAFold(mtx,factor(simuN5$groups))
hist(aFold[[1]])

}

