% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qtotalNormalized}
\alias{qtotalNormalized}
\title{Estimating size factors from the reads count table via ranking}
\usage{
qtotalNormalized(ma, qper = 0.95, qst = 0.85, qstep = 0.01,
  qbound = 0.1, scaleofoutliers = 0.2)
}
\arguments{
\item{ma}{a count matrix}

\item{qper}{quantile for assessing dispersion of data, default is 0.95, which serves to avoid outliers, should in (0,1]}

\item{qst}{start point (quantile) for sliding window, default is 0.85 and should less than 0.99}

\item{qstep}{step of sliding window (quantile), default is 0.01, should at least include one gene per sliding}

\item{qbound}{window size for calculating size factor, default is 0.1, a smaller window size is suitable if number of genes is large}

\item{scaleofoutliers}{baseline scale for detecting outliers of size factor from sliding window, default is 0.2}
}
\value{
a vector with the estimates size factors, one element per column
}
\description{
Function of qtotal for esitmating size factors
}
\details{
Given a matrix of count data, this function esitmates the size
factors by qtotal method, which is based on assessing DE and ranking
}
\examples{

data(simuN5)
counts=simuN5$counts
qtotalNormalized(counts)

}
